+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - all: code format and cleanup
 - python: migrate to logging instead of debug_utils
 - python: move config directory creation from constants to API
+- python/hotkeys (BREAKING): controller naming fix (HotkeysContoller -> HotkeysController)
+- python/hotkeys: hotkey context menu handler refactor
 - as3/components: StatusSwitcher refactor
 - as3/components: fixed component naming (ColorChoise -> ColorChoice)
 - as3/components: fixed internal variables naming (blueSlider) in ColorChoice